requirements.txt
Django>=3.0.0
djangorestframework>=3.12.4
djangorestframework-simplejwt>=4.6.0
django-cors-headers>=3.7.0
psycopg2>=2.8.6
Pillow>=8.2.0
django-storages>=1.11.1
boto3>=1.17.79
whitenoise>=5.2.0
gunicorn>=20.1.0
